+++ src/model/identity_verification_template_reference.rs
@@ -3,7 +3,7 @@ use serde::{Serialize, Deserialize};
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 pub struct IdentityVerificationTemplateReference {
     pub id: String,
-    pub version: f64,
+    pub version: i64,
 }
 impl std::fmt::Display for IdentityVerificationTemplateReference {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
