+++ src/model/pslf_status.rs
@@ -5,9 +5,9 @@ pub struct PslfStatus {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub estimated_eligibility_date: Option<chrono::NaiveDate>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub payments_made: Option<f64>,
+    pub payments_made: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub payments_remaining: Option<f64>,
+    pub payments_remaining: Option<i64>,
 }
 impl std::fmt::Display for PslfStatus {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
